# control-plane/rest/openapi-specs/v0_api_spec.yaml
openapi: 3.0.3
info:
  title: Mayastor RESTful API
  version: v0
servers:
  - url: /v0
paths:
...
  "/nodes/{id}/replicas":
    get:
      tags:
        - Replicas
      operationId: get_node_replicas
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Replica"
        "400":
          description: Request Timeout
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "408":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "412":
          description: Precondition Failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "416":
          description: Range Not satisfiable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "422":
          description: Unprocessable entity
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "501":
          description: Not Implemented
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "503":
          description: Service Unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "504":
          description: Gateway Timeout
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "507":
          description: Insufficient Storage
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
      security:
        - JWT: []
...
  "/nodes/{node_id}/pools/{pool_id}/replicas":
    get:
      tags:
        - Replicas
      operationId: get_node_pool_replicas
      parameters:
        - in: path
          name: node_id
          required: true
          schema:
            type: string
        - in: path
          name: pool_id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Replica"
        "400":
          description: Request Timeout
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "408":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "412":
          description: Precondition Failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "416":
          description: Range Not satisfiable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "422":
          description: Unprocessable entity
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "501":
          description: Not Implemented
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "503":
          description: Service Unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "504":
          description: Gateway Timeout
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "507":
          description: Insufficient Storage
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
      security:
        - JWT: []
  "/nodes/{node_id}/pools/{pool_id}/replicas/{replica_id}":
    get:
      tags:
        - Replicas
      operationId: get_node_pool_replica
      parameters:
        - in: path
          name: node_id
          required: true
          schema:
            type: string
        - in: path
          name: pool_id
          required: true
          schema:
            type: string
        - in: path
          name: replica_id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Replica"
        "400":
          description: Request Timeout
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "408":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "412":
          description: Precondition Failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "416":
          description: Range Not satisfiable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "422":
          description: Unprocessable entity
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "501":
          description: Not Implemented
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "503":
          description: Service Unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "504":
          description: Gateway Timeout
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "507":
          description: Insufficient Storage
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
      security:
        - JWT: []
    put:
      tags:
        - Replicas
      operationId: put_node_pool_replica
      parameters:
        - in: path
          name: node_id
          required: true
          schema:
            type: string
        - in: path
          name: pool_id
          required: true
          schema:
            type: string
        - in: path
          name: replica_id
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateReplicaBody"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Replica"
        "400":
          description: Request Timeout
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "408":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "412":
          description: Precondition Failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "416":
          description: Range Not satisfiable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "422":
          description: Unprocessable entity
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "501":
          description: Not Implemented
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "503":
          description: Service Unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "504":
          description: Gateway Timeout
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "507":
          description: Insufficient Storage
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
      security:
        - JWT: []
    delete:
      tags:
        - Replicas
      operationId: del_node_pool_replica
      parameters:
        - in: path
          name: node_id
          required: true
          schema:
            type: string
        - in: path
          name: pool_id
          required: true
          schema:
            type: string
        - in: path
          name: replica_id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: OK
        "400":
          description: Request Timeout
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "408":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "412":
          description: Precondition Failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "416":
          description: Range Not satisfiable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "422":
          description: Unprocessable entity
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "501":
          description: Not Implemented
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "503":
          description: Service Unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "504":
          description: Gateway Timeout
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "507":
          description: Insufficient Storage
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
      security:
        - JWT: []
  "/nodes/{node_id}/pools/{pool_id}/replicas/{replica_id}/share":
    delete:
      tags:
        - Replicas
      operationId: del_node_pool_replica_share
      parameters:
        - in: path
          name: node_id
          required: true
          schema:
            type: string
        - in: path
          name: pool_id
          required: true
          schema:
            type: string
        - in: path
          name: replica_id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: OK
        "400":
          description: Request Timeout
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "408":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "412":
          description: Precondition Failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "416":
          description: Range Not satisfiable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "422":
          description: Unprocessable entity
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "501":
          description: Not Implemented
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "503":
          description: Service Unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "504":
          description: Gateway Timeout
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "507":
          description: Insufficient Storage
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
      security:
        - JWT: []
  "/nodes/{node_id}/pools/{pool_id}/replicas/{replica_id}/share/{protocol}":
    put:
      tags:
        - Replicas
      operationId: put_node_pool_replica_share
      parameters:
        - in: path
          name: node_id
          required: true
          schema:
            type: string
        - in: path
          name: pool_id
          required: true
          schema:
            type: string
        - in: path
          name: replica_id
          required: true
          schema:
            type: string
            format: uuid
        - in: path
          name: protocol
          required: true
          schema:
            $ref: "#/components/schemas/ReplicaShareProtocol"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: string
        "400":
          description: Request Timeout
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "408":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "412":
          description: Precondition Failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "416":
          description: Range Not satisfiable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "422":
          description: Unprocessable entity
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "501":
          description: Not Implemented
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "503":
          description: Service Unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "504":
          description: Gateway Timeout
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "507":
          description: Insufficient Storage
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
      security:
        - JWT: []
...
  "/pools/{pool_id}/replicas/{replica_id}":
    put:
      tags:
        - Replicas
      operationId: put_pool_replica
      parameters:
        - in: path
          name: pool_id
          required: true
          schema:
            type: string
        - in: path
          name: replica_id
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateReplicaBody"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Replica"
        "400":
          description: Request Timeout
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "408":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "412":
          description: Precondition Failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "416":
          description: Range Not satisfiable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "422":
          description: Unprocessable entity
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "501":
          description: Not Implemented
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "503":
          description: Service Unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "504":
          description: Gateway Timeout
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "507":
          description: Insufficient Storage
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
      security:
        - JWT: []
    delete:
      tags:
        - Replicas
      operationId: del_pool_replica
      parameters:
        - in: path
          name: pool_id
          required: true
          schema:
            type: string
        - in: path
          name: replica_id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: OK
        "400":
          description: Request Timeout
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "408":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "412":
          description: Precondition Failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "416":
          description: Range Not satisfiable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "422":
          description: Unprocessable entity
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "501":
          description: Not Implemented
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "503":
          description: Service Unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "504":
          description: Gateway Timeout
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "507":
          description: Insufficient Storage
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
      security:
        - JWT: []
  "/pools/{pool_id}/replicas/{replica_id}/share":
    delete:
      tags:
        - Replicas
      operationId: del_pool_replica_share
      parameters:
        - in: path
          name: pool_id
          required: true
          schema:
            type: string
        - in: path
          name: replica_id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: OK
        "400":
          description: Request Timeout
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "408":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "412":
          description: Precondition Failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "416":
          description: Range Not satisfiable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "422":
          description: Unprocessable entity
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "501":
          description: Not Implemented
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "503":
          description: Service Unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "504":
          description: Gateway Timeout
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "507":
          description: Insufficient Storage
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
      security:
        - JWT: []
...
components:
  securitySchemes:
    JWT:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    CreateReplicaBody:
      example:
        share: "none"
        size: 80241024
        thin: false
      description: Create Replica Body JSON
      type: object
      properties:
        share:
          $ref: "#/components/schemas/Protocol"
        size:
          description: size of the replica in bytes
          type: integer
          format: int64
          minimum: 0
        thin:
          description: thin provisioning
          type: boolean
      required:
        - share
        - size
        - thin
...
    ReplicaState:
      description: state of the replica
      type: string
      enum:
        - unknown
        - online
        - degraded
        - faulted
    Replica:
      example:
        node: "ksnode-1"
        pool: "pooloop"
        share: "none"
        size: 80241024
        state: Online
        thin: false
        uri: "nvmf://10.1.0.5:8420/nqn.2019-05.io.openebs:fb04022b-1ca1-4789-bcd4-dacbcb54e23c"
        uuid: fb04022b-1ca1-4789-bcd4-dacbcb54e23c
      description: Replica information
      type: object
      properties:
        node:
          description: id of the mayastor instance
          type: string
        pool:
          description: id of the pool
          type: string
        share:
          $ref: "#/components/schemas/Protocol"
        size:
          description: size of the replica in bytes
          type: integer
          format: int64
          minimum: 0
        state:
          $ref: "#/components/schemas/ReplicaState"
        thin:
          description: thin provisioning
          type: boolean
        uri:
          description: uri usable by nexus to access it
          type: string
        uuid:
          description: uuid of the replica
          type: string
          format: uuid
      required:
        - node
        - pool
        - share
        - size
        - state
        - thin
        - uri
        - uuid
    RestJsonError:
      example:
        details: "The Pool 'pooloop' was not found"
        kind: NotFound
      description: Rest Json Error format
      type: object
      properties:
        details:
          description: detailed error information
          type: string
        kind:
          description: error kind
          type: string
          enum:
            - Timeout
            - Deserialize
            - Internal
            - InvalidArgument
            - DeadlineExceeded
            - NotFound
            - AlreadyExists
            - PermissionDenied
            - ResourceExhausted
            - FailedPrecondition
            - NotShared
            - NotPublished
            - AlreadyPublished
            - AlreadyShared
            - Aborted
            - OutOfRange
            - Unimplemented
            - Unavailable
            - Unauthenticated
            - Unauthorized
            - Conflict
            - FailedPersist
            - Deleting
      required:
        - details
        - kind
...
    ReplicaShareProtocol:
      description: Replica Share Protocol
      type: string
      enum:
        - nvmf
    Protocol:
      description: Common Protocol
      type: string
      enum:
        - none
        - nvmf
        - iscsi
        - nbd
...